wattwise/web_application/requirements.txt
streamlit
google-cloud-bigtable
pandas
plotly
